services/web/requirements.txt
Flask==2.3.0
gunicorn==20.1.0
python-whois==0.8.0
dnspython==2.6.1
Werkzeug==2.3.0
dnspython==2.8.0